% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_glm_residuals.R
\name{analyse_glm_residuals}
\alias{analyse_glm_residuals}
\title{This function evaluates residuals from 'glm' model objects. It returns
a normality plot, residuals vs. fitted, and a lag-plot with a trend line
(glm(residual ~ residual -1)) to evaluate autocorrelation. It also
provides a Shapiro-Wilk normality test, Levene test for homogeneity of
variance, a Runs test for randomness, and a t-test for mean zero.}
\usage{
analyse_glm_residuals(glm_object)
}
\arguments{
\item{glm_object}{A glm model object}
}
\value{
Residual plots: normality, residuals vs. fitted data, ACF plot

Residual tests: Mean zero, Shapiro-Wilk normality test, Levene Test for
homogeneity of variance, and Ljung-Box test for autocorrelation
}
\description{
This function evaluates residuals from 'glm' model objects. It returns
a normality plot, residuals vs. fitted, and a lag-plot with a trend line
(glm(residual ~ residual -1)) to evaluate autocorrelation. It also
provides a Shapiro-Wilk normality test, Levene test for homogeneity of
variance, a Runs test for randomness, and a t-test for mean zero.
}
