% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_gls_residuals.R
\name{analyse_gls_residuals}
\alias{analyse_gls_residuals}
\title{This function evaluates residuals from 'gls' model objects generated by the
NLME package. It returns
a normality plot, residuals vs. fitted, and a lag-plot with a trend line
(gls(residual ~ residual -1)) to evaluate autocorrelation. It also
provides a Shapiro-Wilk normality test, Levene test for homogeneity of
variance, a Runs test for randomness, and a t-test for mean zero.}
\usage{
analyse_gls_residuals(gls_object)
}
\arguments{
\item{gls_object}{}
}
\value{
Residual plots: normality, residuals vs. fitted data, autocorrelation
(lagged residuals vs. residuals with fitted linear model).

Residual tests: Mean zero, Shapiro-Wilk normality test, Levene Test for
homogeneity of variance, and Run's test for randomness
}
\description{
This function evaluates residuals from 'gls' model objects generated by the
NLME package. It returns
a normality plot, residuals vs. fitted, and a lag-plot with a trend line
(gls(residual ~ residual -1)) to evaluate autocorrelation. It also
provides a Shapiro-Wilk normality test, Levene test for homogeneity of
variance, a Runs test for randomness, and a t-test for mean zero.
}
