% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatDataframe.R
\name{checkCols}
\alias{checkCols}
\title{check the validation and return index of cols given from input in rawDataFrame}
\usage{
checkCols(rawDataFrame, cols)
}
\arguments{
\item{rawDataFrame}{raw data.frame}

\item{cols}{specific cols given from input}
}
\value{
return validation (only FALSE if invaild cols input) or index of cols
}
\description{
check the validation and return index of cols given from input in rawDataFrame
}
\examples{
library(handyFunctions)
data(people)
checkCols(people, c("..name", "..sex"))
# OR
checkCols(people, c(1, 2))
}
