% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatDataframe.R
\name{checkDtype}
\alias{checkDtype}
\title{Return suggested dtype of vector input}
\usage{
checkDtype(vector)
}
\arguments{
\item{vector}{vector/list input}
}
\value{
Return suggested dtypes of vector
}
\description{
Return suggested dtype of vector input
}
\examples{
library(handyFunctions)
vector <- c(1, 2, 3, "", NA, "  ", "four", "NA", 5)
checkDtype(vector)

}
