% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_numeric.R
\name{choose_numeric}
\alias{choose_numeric}
\title{Selects numberic variables from a data frame}
\usage{
choose_numeric(my_data_frame)
}
\arguments{
\item{data}{}
}
\value{
a tibble
}
\description{
Selects numberic variables from a data frame
}
