% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_characters_to_factors.R
\name{convert_characters_to_factors}
\alias{convert_characters_to_factors}
\title{Convert Characters to Factors}
\usage{
convert_characters_to_factors(data, columns)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{columns}{A character vector specifying the column names
to convert from character to factor.}
}
\value{
The input dataframe with specified columns converted to
factor class.
}
\description{
This function converts specified columns in a dataframe from
character class to factor class.
}
\examples{
data <- data.frame(
  x = c("A", "B", "C"),
  y = c(1, 2, 3),
  z = c("Red", "Green", "Blue")
)
modified_data <- convert_characters_to_factors(data, columns = c("x", "z"))
modified_data

}
