% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_columns_to_numeric.R
\name{convert_columns_to_numeric}
\alias{convert_columns_to_numeric}
\title{Convert Columns to Numeric}
\usage{
convert_columns_to_numeric(data, columns)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{columns}{A character vector specifying the column names
to convert from character to numeric.}
}
\value{
The input dataframe with specified columns converted to
numeric class.
}
\description{
This function converts specified columns in a dataframe
from character class to numeric class, handling problematic
values.
}
\examples{
data <- data.frame(
  x = c("1", "2", "3"),
  y = c("4.5", "6.7", "8.9"),
  z = c("apple", "banana", "orange")
)
modified_data <- convert_columns_to_numeric(data, columns = c("x", "y"))
modified_data

}
