% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_snake_case.R
\name{convert_to_snake_case}
\alias{convert_to_snake_case}
\title{This function replaces spaces with underscores and
converts all strings in a dataframe.to lowercase}
\usage{
convert_to_snake_case(df)
}
\arguments{
\item{df}{The input dataframe.}
}
\value{
A new dataframe with spaces replaced
by underscores and all strings converted to lowercase.
}
\description{
This function replaces spaces with underscores and
converts all strings in a dataframe.to lowercase
}
