% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grade}
\alias{grade}
\title{Grade records of virtual persons in high school}
\format{
A data frame with 6 rows and 7 variables:
\describe{
  \item{name}{name, chinese or foreigner, in carats}
  \item{chinese}{grade of the chinese, in numbers}
  \item{math}{grade of the math, in numbers}
  \item{english}{grade of the english, in numbers}
  \item{physics}{grade of the physics, in numbers}
  \item{biology}{grade of the biology, in numbers}
  \item{chemistry}{grade of the chemistry, in numbers}
  ...

}
}
\source{
"simulated dataset"
}
\usage{
grade
}
\description{
A dataset containing the personal grade information (chinese, math, english, physics, biology, chemistry)
 of virtual persons.
}
\keyword{datasets}
