% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handyFunctions-package.R
\docType{package}
\name{handyFunctions-package}
\alias{handyFunctions}
\alias{handyFunctions-package}
\title{handyFunctions: Useful Functions for Handfully Manipulating and Analyzing Data with Data.frame Format}
\description{
Some useful functions for simply manipulating and analyzing data with data.frame format. It mainly includes the following sections: ReformatDataframe (reformat dataframe with the modifiers), InteractDataframe, and Post-VCF (for downstream analysis for data generated from 'vcftools' Petr et al (2011) \doi{10.1093/bioinformatics/btr330} or 'plink' Chang et al (2015) \doi{10.1186/s13742-015-0047-8}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LuffyLouis/handyFunctions}
  \item Report bugs at \url{https://github.com/LuffyLouis/handyFunctions/issues}
}

}
\keyword{internal}
