% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractDataframe.R
\name{matchIndex}
\alias{matchIndex}
\title{Return the index of source vector matched with query vector}
\usage{
matchIndex(SourceInfo, queryInfo, queryType = TRUE)
}
\arguments{
\item{SourceInfo}{the source vector}

\item{queryInfo}{the query vector}

\item{queryType}{logical If set it to accurate match (default: TRUE)}
}
\value{
the index of source vector matched with query vector
}
\description{
Return the index of source vector matched with query vector
}
\examples{
library(handyFunctions)
data(grade)
matchIndex(grade[, "name"], c("Ming Li", "Bang Wei"))

}
