% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractDataframe.R
\name{mergeCustom}
\alias{mergeCustom}
\title{merge two data.frame based on xcol and ycol}
\usage{
mergeCustom(x, y, xcol, ycol)
}
\arguments{
\item{x}{the first data.frame}

\item{y}{the second data.frame}

\item{xcol}{colnames which you want to merged in first data.frame}

\item{ycol}{colnames which you want to merged in second data.frame}
}
\value{
return the new data.frame merged
}
\description{
merge two data.frame based on xcol and ycol
}
\examples{
library(handyFunctions)
data(people)
data(grade)
mergeCustom(people, grade, "..name", "name")

}
