% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatDataframe.R
\name{modifyRowNames}
\alias{modifyRowNames}
\title{Return reformatted data.frame with standard row names}
\usage{
modifyRowNames(rawDataFrame, rows = TRUE, rawSep = "..", sep = "_")
}
\arguments{
\item{rawDataFrame}{Raw data.frame input}

\item{rows}{Specific row names or indexes what you want to reformat (default: TRUE, use all row)}

\item{rawSep}{Raw odd separation symbol in row names of raw data.frame.
Note: it supports regEx (regular expression), so "." means all possible symbols.
If you want to use the "." dot notation, please use "[.]".}

\item{sep}{Separation symbol in row names of modified data.frame}
}
\value{
A modified data.frame with row names separated by your given delimitator
}
\description{
Return reformatted data.frame with standard row names
}
\examples{
library(handyFunctions)
data(people)
modifyRowNames(people)

}
