% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people}
\alias{people}
\title{Basic information of virtual persons}
\format{
A data frame with 6 rows and 4 variables:
\describe{
  \item{..name}{name, chinese or foreigner, in carats}
  \item{..sex}{sex of the person, in carats}
  \item{..age}{living age in final record, in numbers}
  \item{..death..age}{final age when a person is dead, in numbers}
  ...

}
}
\source{
"simulated dataset"
}
\usage{
people
}
\description{
A dataset containing the personal basic information (name, sex, age, and death_age) of virtual persons.
}
\keyword{datasets}
