% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractDataframe.R
\name{queryingInfo}
\alias{queryingInfo}
\title{return index of x data.frame with the given vector/list or ycol in data.frame
(if set the accurate match or not)}
\usage{
queryingInfo(SourceData, sourceCol, queryCol, queryInfo, queryType = TRUE)
}
\arguments{
\item{SourceData}{the source data.frame which you want to query}

\item{sourceCol}{the col names or index of query field in source data.frame}

\item{queryCol}{the col names or index of return field in source data.frame}

\item{queryInfo}{vector/list the query info}

\item{queryType}{logical if set it to accurate match (default: TRUE)}
}
\value{
a vector in query field matched with query info in source data
}
\description{
return index of x data.frame with the given vector/list or ycol in data.frame
(if set the accurate match or not)
}
\examples{
library(handyFunctions)
data(grade)
queryingInfo(grade, "name", "chinese", c("Ming Li", "Bang Wei"))

}
