% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename Columns Based on a List}
\usage{
rename_columns(data, new_names)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{new_names}{A character vector specifying the new
column names in the desired order.}
}
\value{
The input dataframe with renamed columns.
}
\description{
This function renames multiple columns in a dataframe based
on a list of new names.
}
\examples{
data <- data.frame(
  col1 = c(1, 2, 3),
  col2 = c(4, 5, 6),
  col3 = c(7, 8, 9)
)
new_names <- c("new_col1", "new_col2", "new_col3")
renamed_data <- rename_columns(data, new_names)
renamed_data

}
