% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_columns.R
\name{select_columns}
\alias{select_columns}
\title{Select Columns from a Dataframe}
\usage{
select_columns(data, column_names)
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{column_names}{A character vector specifying the column
names to be selected.}
}
\value{
A new dataframe with only the selected columns.
}
\description{
This function selects specified columns from a dataframe and
returns a new dataframe with only the selected columns.
}
\examples{
data <- data.frame(
  x = c(1, 2, 3),
  y = c(4, 5, 6),
  z = c(7, 8, 9)
)
selected_data <- select_columns(data, column_names = c("x", "z"))
selected_data

}
