% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SNV_1MB_density_data}
\alias{SNV_1MB_density_data}
\title{The SNPV number within 1Mb bins at chromosome levels generated from transcriptome dataset of two dog populations
(including wild wolf and domesticated dogs).}
\format{
A data frame with 2544 rows and 4 variables:
\describe{
  \item{CHROM}{chrom id, reference genome of CanFam3.1, in numbers/carats}
  \item{BIN_START}{the start genomic coordinate for one bin at relevant chromosome, in numbers}
  \item{SNP_COUNT}{the end genomic coordinate for one bin at relevant chromosome, in numbers}
  \item{VARIANTS.KB}{SNV(variants) number within one bin per KB, in numbers}
  ...

}
}
\source{
"real dataset"
}
\usage{
SNV_1MB_density_data
}
\description{
A dataset containing the SNV number within 1Mb bins called from transcriptome dataset of wild wolf and domesticated dogs.
}
\keyword{datasets}
