% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReformatDataframe.R
\name{splitCol}
\alias{splitCol}
\title{Return specific-indexed vector according to given delimitator/separator by splitting one col in data.frame}
\usage{
splitCol(data, col = FALSE, sep, index, fixed = TRUE)
}
\arguments{
\item{data}{vector or data.frame input}

\item{col}{the col names or indexes if data.frame input}

\item{sep}{separation deliminator}

\item{index}{the index of symbol which you want}

\item{fixed}{logical. If TRUE match split exactly, otherwise use regular expressions, detailed info can be seen in \href{/library/base/help/strsplit}{strsplit}.}
}
\value{
specific-indexed vector or factor
}
\description{
Return specific-indexed vector according to given delimitator/separator by splitting one col in data.frame
}
\examples{
library(handyFunctions)
data(people)
splitCol(people, col = 1, sep = " ", index = 2)

}
