% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{symmetrize}
\alias{symmetrize}
\title{symmetrize}
\usage{
symmetrize(your.dt)
}
\arguments{
\item{your.dt}{a data.table generated from the txt file exported by Gremlin.
Must have columns 'i', 'j', 'i_id', 'j_id', and 'prob'}
}
\value{
a data.table that is now includes a symmetric list of contacts
}
\description{
symmetrize
}
\examples{
a <- symmetrize(dt)
}
