% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_norm.R
\name{tests_norm}
\alias{tests_norm}
\title{Tests normality}
\usage{
tests_norm(iris, Sepal.Length)
}
\arguments{
\item{df}{A dataframe}

\item{x}{A dataframe column with numeric values}
}
\value{
A plot grid with normality plot, histogram, density plot, and
histogram with normality curve. Prints Shapiro-Wilke test results.
}
\description{
Performs Shapiro-Wilk test and makes normality plot (qq), density plot,
histogram, and histogram with normality curve.
}
