% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_variance.R
\name{tests_variance}
\alias{tests_variance}
\title{Tests Variance}
\usage{
tests_variance(iris, Sepal.Length, Species)
}
\arguments{
\item{df}{A dataframe}

\item{data_var}{A dataframe column with numeric values}

\item{group_var}{A dataframe column with a grouping variable (factor)}
}
\value{
A plot grid with boxplot, boxplot with dots, and density plot.
Prints Levene test results.
}
\description{
Performs Levene tests and makes boxplots and density plot
}
