% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handyFunctions-package.R
\docType{package}
\name{handyFunctions-package}
\alias{handyFunctions}
\alias{handyFunctions-package}
\title{handyFunctions: Useful functions for handfully manipulating and analyzing data with data.frame format}
\description{
Some useful functions for simply manipulating and analyzing data with data.frame format. It mainly includes the following sections: ReformatDataframe modifyColnames: modify the colnames of raw data.frame modifyColTypes: automatically recognize and change the data types of each col modifyRownames: modify the rownames of raw data.frame unifyDataframe: reformat dataframe with the above all modifiers simultaneously splitCol: return specific-indexed vector according to given delimitator/separator by splitting one col in data.frame InteractDataframe mergeCustom: merge two data.frame based on xcol and ycol matchIndex: return index of x data.frame with the given vector/list or ycol in data.frame (if set the accurate match or not) Post-VCF plotSNVdensity: Function of showing SNP density at chromosome level
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LuffyLouis/handyFunctions}
  \item Report bugs at \url{https://github.com/LuffyLouis/handyFunctions/issues}
}

}
\keyword{internal}
