% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handyFunctions-package.R
\docType{package}
\name{handyFunctions-package}
\alias{handyFunctions}
\alias{handyFunctions-package}
\title{handyFunctions: Useful functions for handfully manipulating and analyzing data with data.frame format}
\description{
Some useful functions for simply manipulating and analyzing data with data.frame format. It mainly includes the following sections: ReformatDataframe (reformat dataframe with the modifiers), InteractDataframe, and Post-VCF (for downstream analysis for data generated from vcftools or plink).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LuffyLouis/handyFunctions}
  \item Report bugs at \url{https://github.com/LuffyLouis/handyFunctions/issues}
}

}
\keyword{internal}
