% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ggplot.R
\name{save_ggplot}
\alias{save_ggplot}
\title{Saves ggplot objects as .svg at a high resolution}
\usage{
save_ggplot(
  filename,
  plot_object,
  path,
  width = 20,
  height = 20,
  units = "cm",
  dpi = 300
)
}
\arguments{
\item{filename}{name for the file that will be saved}

\item{plot_object}{a ggplot object}

\item{path}{filepath to save file}
}
\value{
saves ggplot in chosen location
}
\description{
Saves ggplot objects as .svg at a high resolution
}
