% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ggplot.R
\name{save_ggplot}
\alias{save_ggplot}
\title{Save a ggplot Object as an SVG File at High Resolution}
\usage{
save_ggplot(
  plot_object,
  filename,
  path,
  width = 20,
  height = 20,
  units = "cm",
  dpi = 300
)
}
\arguments{
\item{plot_object}{A \code{ggplot} object to be saved.}

\item{filename}{The name of the file in which the plot will be saved. The name should end with the \code{.svg} extension.}

\item{path}{The directory path where the file will be saved.}

\item{width}{The width of the saved plot (default is 20).}

\item{height}{The height of the saved plot (default is 20).}

\item{units}{The units for the width and height (default is "cm"). Common options include "in", "cm", and "mm".}

\item{dpi}{The resolution of the saved plot in dots-per-inch (default is 300).}
}
\value{
None. The function saves the plot to the specified file.
}
\description{
This function saves a \code{ggplot} object to a specified directory as a high-resolution SVG file.
}
\examples{
\dontrun{
library(ggplot2)
plot_object <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
save_ggplot(plot_object, "example_plot.svg", tempdir())
}
}
